+++ client/cody/src/chat/ChatViewProvider.ts
@@ -17,7 +17,7 @@ import { SourcegraphEmbeddingsSearchClient } from '@sourcegraph/cody-shared/src/
 import { Guardrails, annotateAttribution } from '@sourcegraph/cody-shared/src/guardrails'
 import { highlightTokens } from '@sourcegraph/cody-shared/src/hallucinations-detector'
 import { IntentDetector } from '@sourcegraph/cody-shared/src/intent-detector'
-import { MAX_AVAILABLE_PROMPT_LENGTH, SOLUTION_TOKEN_LENGTH } from '@sourcegraph/cody-shared/src/prompt/constants'
+import { ANSWER_TOKENS, DEFAULT_MAX_TOKENS } from '@sourcegraph/cody-shared/src/prompt/constants'
 import { Message } from '@sourcegraph/cody-shared/src/sourcegraph-api'
 import { SourcegraphGraphQLAPIClient } from '@sourcegraph/cody-shared/src/sourcegraph-api/graphql'
 import { isError } from '@sourcegraph/cody-shared/src/utils'
@@ -63,6 +63,19 @@ export type Config = Pick<
     | 'experimentalGuardrails'
 >
 
+/**
+ * The problem with a token limit for the prompt is that we can only
+ * estimate tokens (and do so in a very cheap way), so it can be that
+ * we undercount tokens. If we exceed the maximum tokens, things will
+ * start to break, so we should have some safety cushion for when we're wrong in estimating.
+ *
+ * Ie.: Long text, 10000 characters, we estimate it to be 2500 tokens.
+ * That would fit into a limit of 3000 tokens easily. Now, it's actually
+ * 3500 tokens, because it splits weird and our estimation is off, it will
+ * fail. That's where we want to add this safety cushion in.
+ */
+const SAFETY_PROMPT_TOKENS = 100
+
 export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disposable, IdleRecipeRunner {
     private isMessageInProgress = false
     private cancelCompletionCallback: (() => void) | null = null
@@ -74,7 +87,6 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
     private inputHistory: string[] = []
     private chatHistory: ChatHistory = {}
 
-    private maxPromptLimit = MAX_AVAILABLE_PROMPT_LENGTH
     private transcript: Transcript = new Transcript()
 
     // Allows recipes to hook up subscribers to process sub-streams of bot output
@@ -118,12 +130,6 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
             }),
             vscode.commands.registerCommand('cody.auth.sync', () => this.syncAuthStatus())
         )
-        const codyConfig = vscode.workspace.getConfiguration('cody')
-        const tokenLimit = codyConfig.get<number>('provider.limit.prompt')
-        const solutionLimit = codyConfig.get<number>('provider.limit.solution') || SOLUTION_TOKEN_LENGTH
-        if (tokenLimit) {
-            this.maxPromptLimit = tokenLimit - solutionLimit
-        }
     }
 
     private idleCallbacks_: (() => void)[] = []
@@ -515,7 +521,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
 
                 const { prompt, contextFiles } = await this.transcript.getPromptForLastInteraction(
                     getPreamble(this.codebaseContext.getCodebase()),
-                    this.maxPromptLimit
+                    this.maxPromptTokens
                 )
                 this.transcript.setUsedContextFilesForLastInteraction(contextFiles)
                 this.sendPrompt(prompt, interaction.getAssistantMessage().prefix ?? '')
@@ -548,7 +554,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
 
         const { prompt, contextFiles } = await transcript.getPromptForLastInteraction(
             getPreamble(this.codebaseContext.getCodebase()),
-            this.maxPromptLimit
+            this.maxPromptTokens
         )
         transcript.setUsedContextFilesForLastInteraction(contextFiles)
 
@@ -912,6 +918,27 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
             disposable.dispose()
         }
         this.disposables = []
+    }
+
+    private get maxPromptTokens(): number {
+        const authStatus = this.authProvider.getAuthStatus()
+
+        const codyConfig = vscode.workspace.getConfiguration('cody')
+        const tokenLimit = codyConfig.get<number>('provider.limit.prompt')
+        const localSolutionLimit = codyConfig.get<number>('provider.limit.solution')
+
+        // The local config takes precedence over the server config.
+        if (tokenLimit && localSolutionLimit) {
+            return tokenLimit - localSolutionLimit
+        }
+
+        const solutionLimit = (localSolutionLimit || ANSWER_TOKENS) + SAFETY_PROMPT_TOKENS
+
+        if (authStatus.configOverwrites?.chatModelMaxTokens) {
+            return authStatus.configOverwrites.chatModelMaxTokens - solutionLimit
+        }
+
+        return DEFAULT_MAX_TOKENS - solutionLimit
     }
 }