+++ client/jetbrains/src/main/java/com/sourcegraph/cody/completions/prompt_library/WebviewErrorMessenger.java
@@ -1,7 +1,11 @@
 package com.sourcegraph.cody.completions.prompt_library;
 
+import com.intellij.openapi.diagnostic.Logger;
+
 public class WebviewErrorMessenger {
+  private static final Logger logger = Logger.getInstance(WebviewErrorMessenger.class);
+
   public void show(String message) {
-    System.out.println("WebviewErrorMessenger: " + message);
+    logger.error("WebviewErrorMessenger: " + message);
   }
 }