+++ cmd/migrator/generate.sh
@@ -76,9 +76,8 @@ git_versions=(
   v4.4.0 v4.4.1 v4.4.2
   v4.5.0 v4.5.1
   v5.0.0 v5.0.1 v5.0.2 v5.0.3 v5.0.4
- v5.1.0 v5.1.1)
+  v5.1.0 v5.1.1 v5.1.2)
 
 for version in "${git_versions[@]}"; do
   download_github "${version}"
 done
-