+++ enterprise/cmd/cody-gateway/internal/actor/productsubscription/productsubscription.go
@@ -178,7 +178,10 @@ func (s *Source) checkAccessToken(ctx context.Context, token string) (*dotcom.Ch
 
 	for _, gqlerr := range gqlerrs {
 		if gqlerr.Extensions != nil && gqlerr.Extensions["code"] == codygateway.GQLErrCodeProductSubscriptionNotFound {
-			return nil, actor.ErrAccessTokenDenied{Reason: "associated product subscription not found"}
+			return nil, actor.ErrAccessTokenDenied{
+				Source: s.Name(),
+				Reason: "associated product subscription not found",
+			}
 		}
 	}
 	return nil, err