+++ enterprise/cmd/embeddings/shared/main_test.go
@@ -335,5 +335,4 @@ func TestEmbeddingModelMismatch(t *testing.T) {
 			}
 		})
 	}
-
 }